% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USA_state_boundaries.R
\docType{package}
\name{USA.state.boundaries-package}
\alias{USA.state.boundaries}
\alias{USA.state.boundaries-package}
\title{USA.state.boundaries: Replacement of the 'USGSstates2k' Package}
\description{
This is a replacement of the archived 'USGSstates2k' package. It contains a map with the state boundaries of the USA obtained from the USGS. Irucka worked with this data set while a Cherokee Nation Technology Solutions (CNTS) USGS Contractor and/or USGS employee. This package contains a subset of the data included in the 'USA.state.boundaries.data' package, which is available in a 'drat' repository. To install that data package, please follow the instructions at <https://gitlab.com/iembry/usa.state.boundaries.data>.
}
\details{
USA.state.boundaries: Replacement of the 'USGSstates2k' Package

This package contains a subset of the data included in the
`USA.state.boundaries.data` package, which is available in a `drat`
repository. To access that data package, please read this package's
README and refer to this package's vignette.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/iembry/usa.state.boundaries}
  \item Report bugs at \url{https://gitlab.com/iembry/usa.state.boundaries/-/issues}
}

}
\keyword{internal}
