% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USA_state_boundaries.R
\docType{package}
\name{USA.state.boundaries-package}
\alias{USA.state.boundaries}
\alias{USA.state.boundaries-package}
\title{USA.state.boundaries: WGS84 Datum Map of the USA, Including Puerto Rico and the U.S. Virgin Islands}
\description{
Contains a WGS84 datum map of the USA, which includes all Commonwealth and State boundaries & also includes Puerto Rico and the U.S. Virgin Islands. This map is a reprojection of the NAD83 datum map from the USGS National Map. This package contains a subset of the data included in the 'USA.state.boundaries.data' package, which is available in a 'drat' repository. To install that data package, please follow the instructions at \url{https://gitlab.com/iembry/usa.state.boundaries.data}.
}
\details{
USA.state.boundaries: Replacement of the 'USGSstates2k' Package

This package contains a subset of the data included in the
`USA.state.boundaries.data` package, which is available in a `drat`
repository. To access that data package, please read this package's
README.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/iembry/usa.state.boundaries}
  \item Report bugs at \url{https://gitlab.com/iembry/usa.state.boundaries/-/issues}
}

}
\keyword{internal}
